backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.0
pydantic>=2.7.4
asyncpg==0.29.0
httpx==0.27.0
redis==5.0.8
python-ulid==2.2.0
psycopg2-binary==2.9.9
openpyxl==3.1.2
langchain-aws==0.2.14
langchain-core>=0.3.27
